Move MerkleRadixStore to store module
This change adds a store submodule as the new home for the
MerkleRadixStore and implementations.

It moves the MerkleRadixStore trait, and the base SqlMerkleRadixStore to
this new module. The references to these types are updated accordingly.
